requirements.txt
djangorestframework==3.13.1
django-cors-headers==3.13.0
django==3.2.13
django-environ==0.9.0
pandas==1.4.3
numpy==1.23.0
openpyxl==3.0.10

Pillow==9.2.0
django-smuggler==1.0.4
environ==1.0
djangorestframework-simplejwt==5.2.0